piggy_bank_bot/requirements.txt
aiogram>=3.0.0
supabase>=2.0.0
python-dotenv>=1.0.0